pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "the-claw-renovate-showcase"
version = "0.1.0"
description = "A small Renovate PR showcase for an external dependency triage agent."
requires-python = ">=3.10"
dependencies = [
    "pydantic==1.10.13",
    "requests==2.28.0",
    "httpx==0.28.1",
    "click==8.1.0",
    "python-dateutil==2.9.0.post0",
    "pyyaml==6.0",
    "rich==13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
